pyproject.toml
[project]
name = "rtt_display"
version = "0.1.0"
description = "Display realtime train depature information from any UK station using an SSD1322 OLED display and a Raspberry Pi zero."
authors = [{name = "Sam Sims", email = "samsimss98@gmail.com"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.3",
    "luma-core>=2.4.2",
    "luma-emulator>=1.5.0",
    "luma-oled>=3.13.0",
    "types-requests>=2.32.0.20240712",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11.2",
]

[project.scripts]
rtd = "rtt_display.display:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # pathlib
    "PTH",
]

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"
